IOL ORDER ROUTER
Former-commit-id: 191bccb943b8d139c313ad6169857df48c559958

+++ HFTPlatform/zHFT.OrderRouters.InvertirOnline.Common/Responses/Trade.cs
@@ -12,7 +12,7 @@ namespace zHFT.OrderRouters.InvertirOnline.Common.Responses
 
         public DateTime? fecha { get; set; }
 
-        public int cantidad { get; set; }
+        public double cantidad { get; set; }
 
         public double precio { get; set; }
